+++ src/client/index.ts
@@ -1,7 +1,12 @@
-import { UnpackArray } from '../utils'
-import type { GetGamesBySlotQuery } from './graphql'
+import { ContentsOf, UnpackArray } from '../utils'
+import type { GetGamesBySlotQuery, GetScheduleQuery } from './graphql'
 
 export * from './graphql'
 
 export type GameArray = NonNullable<GetGamesBySlotQuery['games']>['edges']
 export type GameEntry = NonNullable<Omit<UnpackArray<GameArray>, '__typename'>['node']>
+
+export type GameAssignmentNode = Omit<
+  Omit<ContentsOf<ContentsOf<GetScheduleQuery, 'gameAssignments'>, 'nodes'>, '__typename'>,
+  'nodeId'
+> & { nodeId?: string }